+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - [reactor package] reactor design mode for resolving sync locking and concurrent controlling.
 - [chanpool package] do select for multiple channels which is ambiguous
 - [trigger package] pack input elements stream to batch by controlling params or function
+- [values package] strings matcher with multiple regex, wildcard pre-built 
 ---
 # [v0.9.0] 2024-01-12
 ## Features